feat(console): move listAssets
