+++ src/console/src/storage/store.rs
@@ -1,4 +1,3 @@
-use crate::memory::STATE;
 use crate::storage::certified_assets::runtime::init_certified_assets as init_runtime_certified_assets;
 use crate::storage::state::heap::{
     delete_domain, get_config, get_domain, get_domains, insert_config, insert_domain,
@@ -8,52 +7,15 @@ use crate::store::stable::get_proposal;
 use crate::strategies_impls::storage::StorageState;
 use candid::Principal;
 use junobuild_cdn::proposals::ProposalId;
-use junobuild_collections::types::core::CollectionKey;
-use junobuild_shared::list::list_values;
 use junobuild_shared::types::core::DomainName;
 use junobuild_shared::types::domain::CustomDomains;
-use junobuild_shared::types::list::{ListParams, ListResults};
-use junobuild_storage::heap_utils::collect_assets_heap;
 use junobuild_storage::store::create_batch;
 use junobuild_storage::types::config::StorageConfig;
-use junobuild_storage::types::interface::{AssetNoContent, InitAssetKey};
+use junobuild_storage::types::interface::InitAssetKey;
 use junobuild_storage::types::runtime_state::BatchId;
-use junobuild_storage::types::state::FullPath;
-use junobuild_storage::types::store::Asset;
-use junobuild_storage::utils::{map_asset_no_content};
 use junobuild_storage::well_known::update::update_custom_domains_asset;
 use junobuild_storage::well_known::utils::build_custom_domain;
 use regex::Regex;
-
-pub fn list_assets(
-    collection: &CollectionKey,
-    filters: &ListParams,
-) -> Result<ListResults<AssetNoContent>, String> {
-    STATE.with(|state| {
-        let state_ref = state.borrow();
-        let assets = collect_assets_heap(collection, &state_ref.heap.storage.assets);
-        Ok(list_assets_impl(&assets, filters))
-    })
-}
-
-fn list_assets_impl(
-    assets: &[(&FullPath, &Asset)],
-    filters: &ListParams,
-) -> ListResults<AssetNoContent> {
-    let values = list_values(assets, filters);
-
-    ListResults::<AssetNoContent> {
-        items: values
-            .items
-            .into_iter()
-            .map(|(_, asset)| map_asset_no_content(&asset))
-            .collect(),
-        items_length: values.items_length,
-        items_page: values.items_page,
-        matches_length: values.matches_length,
-        matches_pages: values.matches_pages,
-    }
-}
 
 pub fn init_asset_upload(
     caller: Principal,
