+++ src/libs/cdn/src/storage/state/heap/mod.rs
@@ -3,3 +3,4 @@ pub mod store;
 
 pub use state::*;
 pub use store::get_public_asset;
+pub use store::list_assets;
